refactor(broker): split offchain_market_monitor.rs into directory layout (C39 of PR-3)
Pure mechanical move. offchain_market_monitor.rs (130 lines) becomes:

- offchain_market_monitor/mod.rs (~27 lines): module decls + pub use
  service::OffchainMarketMonitor.
- offchain_market_monitor/error.rs (~38 lines): OffchainMarketMonitorErr
  enum + coded_error_impl!.
- offchain_market_monitor/service.rs (~111 lines): OffchainMarketMonitor
  struct + impl + BrokerService impl.

External path crate::offchain_market_monitor::OffchainMarketMonitor
preserved via the re-export.
